-计划修改车次可以为小数

查看合并请求 root/Auxiliary!117

+++ plan/serializers.py
@@ -243,7 +243,7 @@ class DownRegulationSerializer(BaseModelSerializer):
 
 class UpdateTrainsSerializer(BaseModelSerializer):
     '''修改车次和重传'''
-    trains = serializers.IntegerField(write_only=True, help_text='修改车次')
+    trains = serializers.DecimalField(write_only=True, help_text='修改车次',decimal_places=1, max_digits=8)
 
     class Meta:
         model = ProductClassesPlan
@@ -252,7 +252,7 @@ class UpdateTrainsSerializer(BaseModelSerializer):
 
     @atomic()
     def update(self, instance, validated_data):
-        if int(validated_data.get('trains')) - int(instance.plan_trains) <= 2:
+        if validated_data.get('trains') - instance.plan_trains <= 2:
             raise serializers.ValidationError({'trains': "修改车次至少要比原车次大2次"})
         p_obj = PlanStatus.objects.filter(plan_classes_uid=instance.plan_classes_uid).order_by('created_date').last()
         if not p_obj:
